Create OutputDiv component which shows state when the button is pressed
currently it's using the deprecated "componentWillReceiveProps" though
so I intend to update it. also, would like to have the cell click
trigger the OutputDiv update, but i'm not entirely sure how to do that
just yet. (I'd guess either an event emitter or a click handler which
gets passed down to Cell)